test(connectors): inline the real-payload avro regression test into kafka.rs
Per the repo convention called out in #50 review ("we don't have any standalone
test files in the streamling-core crate"), move arrow_avro_equivalence.rs into
an inline #[cfg(test)] module next to the Kafka source that owns the decode
path. Same fixtures (crate-local testdata/), same assertions; the connectors
crate no longer carries a tests/ directory.
